% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentProcessedLevel.R
\name{currentProcessedLevel}
\alias{currentProcessedLevel}
\title{currentProcessedLevel}
\usage{
currentProcessedLevel(dbName, table, subject, channel, suid, timestamp)
}
\description{
currentProcessedLevel
}
\examples{
currentProcessedLevel()
}
