% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContextFromTaskTable.R
\name{getContextFromTaskTable}
\alias{getContextFromTaskTable}
\title{getContextFromTaskTable}
\usage{
getContextFromTaskTable(conn, ...)
}
\description{
Provides default vaules for project and database use based on the previous values.
}
\examples{
getContextFromTaskTable()
}
