% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markAsProcessed.R
\name{markAsProcessed}
\alias{markAsProcessed}
\title{markAsProcessed}
\usage{
markAsProcessed(dbName, table, subject, channel, suid, timestamp, flag)
}
\description{
markAsProcessed
}
\examples{
case <- list()
case['subject'] <- "testSubject"
case['channel'] <- "testChannel"
case['event_start'] <- 0
conn <- db( 'testProject' )
markAsProcessed( conn, case, flag=1 )
}
