% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markProcessedLevel.R
\name{markProcessedLevel}
\alias{markProcessedLevel}
\title{markProcessedLevel}
\usage{
markProcessedLevel(compArgs, case, newProcessLevel)
}
\description{
markProcessedLevel
}
\examples{
case <- list()
case['subject'] <- "testSubject"
case['channel'] <- "testChannel"
case['event_start'] <- 0
conn <- db( 'testProject' )
markAsProcessed( conn, case, flag=1 )
}
