% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validContext.R
\name{validContext}
\alias{validContext}
\title{getContextFromTaskTable}
\usage{
validContext(context)
}
\description{
Provides default vaules for project and database use based on the previous values.
}
\examples{
getContextFromTaskTable()
}
